update  (#1)
* remove info output that slow down scene loading

* address remap issue

* add basic instance support by repeating geometry

* add proper offset based on PBRT code

* fix uncorrectly oriented normals

* try to found workaround robust intersection (cornel-box)

* remove geometry changes for now

* point-normal project

* clean scripts for point-normal (master branch)

* rename point-normal files

* update dependencies

* update dependencies, fix build

* update web app dependencies - generates compilation errors

* update rust -- compiler issues

* minimal version with path tracing working

* finish fixing arg parser

* reformat

---------

+++ src/integrators/direct.rs
@@ -3,8 +3,8 @@ use crate::integrators::*;
 use cgmath::InnerSpace;
 
 pub struct IntegratorDirect {
-    pub nb_bsdf_samples: u32,
-    pub nb_light_samples: u32,
+    pub nb_bsdf_samples: usize,
+    pub nb_light_samples: usize,
 }
 
 impl Integrator for IntegratorDirect {
